New Menu, NewFunc
Added a new menu for various actions.
Added alternate functionality to the choose-root button to allow the
user to select a file instead of a folder.
Added handling to the Find button to correctly set the slashes in the
text fields and then kick off the search via a call to DataFactory.