Search: perform, folder, file implementation
Added code for checking if a folder matches the search parameters.
Added code for checking the contents of a folder.
Added code for checking if a file matches the search parameters.
Added code for checking the contents of a file.
Added code for printing the current search path.
Added code to guard against reading the search's output file.