ModFunc: SessionFolder; Rename: JarFile
Changed the SessionFolder getter so that it makes the folder exist (thus
session files can be listed without causing a file-doesn't-exist
exception).
Renamed JarFile to JarLocationFile to make it more indicative of the
fact that it returns the file where the Jar's location is stashed,
rather than the Jar filepath itself.